fix some bugs in test for getparents

+++ src/sf_map.h
@@ -168,7 +168,7 @@ class SFMap {
          *  e.g. result[i] = the node before the ith node on the shortest path between start and
          *  the ith nodes.
          */
-        vector<double> getParents(int start) const;
+        vector<int> getParents(int start) const;
 
         /**
          * For testing
